move get_custom_fields_strict() to cf mixin
