+++ app/extensions/__init__.py
@@ -589,6 +589,22 @@ class CustomFieldMixin(object):
 
     # TODO we probably want one to ADD values to a multiple=TRUE list-type
 
+    def get_custom_fields_strict(self):
+        from app.modules.site_settings.helpers import SiteSettingCustomFields
+
+        cf = self.custom_fields
+        cf_strict = {}
+        for cfd_id in cf:
+            if not SiteSettingCustomFields.is_valid_value_for_class(
+                self.__class__.__name__, cfd_id, cf[cfd_id]
+            ):
+                log.warning(
+                    f'skipping custom field definition {cfd_id} with invalid value "{cf[cfd_id]}" on {self}'
+                )
+            else:
+                cf_strict[cfd_id] = cf[cfd_id]
+        return cf_strict
+
 
 ##########################################################################################
