+++ app/modules/sightings/models.py
@@ -318,22 +318,6 @@ class Sighting(db.Model, HoustonModel, CustomFieldMixin):
     def get_custom_fields(self):
         return self.custom_fields
 
-    def get_custom_fields_strict(self):
-        from app.modules.site_settings.helpers import SiteSettingCustomFields
-
-        cf = self.custom_fields
-        cf_strict = {}
-        for cfd_id in cf:
-            if not SiteSettingCustomFields.is_valid_value_for_class(
-                'Sighting', cfd_id, cf[cfd_id]
-            ):
-                log.warning(
-                    f'skipping custom field definition {cfd_id} with invalid value "{cf[cfd_id]}" on {self}'
-                )
-            else:
-                cf_strict[cfd_id] = cf[cfd_id]
-        return cf_strict
-
     def init_progress_identification(self, overwrite=False):
         from app.modules.progress.models import Progress
